// packages/tldraw/packages/core/src/types/TLCursor.ts
export enum TLCursor {
  None = 'none',
  Default = 'default',
  Pointer = 'pointer',
  Cross = 'crosshair',
  Grab = 'grab',
  Rotate = 'rotate',
  Wait = 'wait',
  Progress = 'progress',
  Grabbing = 'grabbing',
  ResizeEdge = 'resize-edge',
  ResizeCorner = 'resize-corner',
  Text = 'text',
  Move = 'move',
  EwResize = 'ew-resize',
  NsResize = 'ns-resize',
  NeswResize = 'nesw-resize',
  NwseResize = 'nwse-resize',
  NeswRotate = 'nesw-rotate',
  NwseRotate = 'nwse-rotate',
  SwneRotate = 'swne-rotate',
  SenwRotate = 'senw-rotate',
}